refactor: replace hardcoded color with Tailwind class
Update color to Tailwind's slate-75 for better maintainability.

#VERCEL_SKIP
