+++ components/market-filter/index.ts
@@ -1,2 +1,2 @@
-export { default as MarketFilterDropdown } from "./market-filter-dropdown"
-export type { MarketFilterOptions } from "./market-filter-dropdown"
+export { default as MarketFilterDropdown } from "./market-filter-component"
+export type { MarketFilterOptions } from "./market-filter-component"
